+++ mycodo/inputs/examples/k30_ttn.py
@@ -76,17 +76,17 @@ class InputModule(AbstractInput):
 
     def __init__(self, input_dev, testing=False):
         super(InputModule, self).__init__(input_dev, testing=testing, name=__name__)
+
         self.timer = 0
+
+        # custom options
+        self.serial_device = None
+        # set custom_options
+        self.setup_custom_options(
+            INPUT_INFORMATION['custom_options'], input_dev)
 
         if not testing:
             import serial
-
-            if input_dev.custom_options:
-                for each_option in input_dev.custom_options.split(';'):
-                    option = each_option.split(',')[0]
-                    value = each_option.split(',')[1]
-                    if option == 'serial_device':
-                        self.serial_device = value
 
             self.uart_location = input_dev.uart_location
             self.baud_rate = input_dev.baud_rate