+++ mycodo/inputs/sht31_smart_gadget.py
@@ -138,6 +138,7 @@ class InputModule(AbstractInput):
         self.download_stored_data = None
         self.logging_interval_ms = None
         self.logging_interval = None
+        # set custom_options
         self.setup_custom_options(
             INPUT_INFORMATION['custom_options'], input_dev)
         self.logging_interval_ms = self.logging_interval * 1000